requirements.txt
python-dotenv
semantic-kernel
azure-identity